More experimentation with optimizing candidates. Add ability to export SFT data in text format. Add options to specify parameters for elliptic injections.

+++ lalapps/src/pulsar/PowerFlux/lines.c
@@ -7,7 +7,7 @@
 #include "lines.h"
 #include "cmdline.h"
 
-extern long nbins, first_bin;
+extern int nbins, first_bin;
 extern FILE *LOG;
 
 #if 0
